requirements_fixed.txt
torch
torchaudio
speechbrain==1.0.2
openai-whisper==20231117
yt-dlp==2024.10.22
moviepy==2.0.0.dev2
imageio==2.35.1
imageio-ffmpeg==0.5.1
transformers==4.44.2
sentencepiece==0.2.0
noisereduce==3.0.2
streamlit==1.39.0
numpy==1.26.4
librosa==0.10.2.post1
pydub==0.25.1
scipy==1.14.1
